v0.1.5: admin frontend MVP complete
- Frontend: Vue3 admin panel fully functional (15/15 pages pass)
- Backend: API Key encryption (AES-GCM-256), auth middleware fixes
- Docker: WSL2 deployment working
- USDT: wallet table migration added
- Version bump: 0.1.1 -> 0.1.5
- Cleanup: remove temp scripts, ui/ renamed to frontend/

+++ rsws_db/src/wallet.rs
@@ -6,6 +6,7 @@ use chrono::{DateTime, Utc};
 use rsws_common::error::RswsError;
 use rsws_common::snowflake;
 use serde::{Deserialize, Serialize};
+use sqlx::types::Decimal;
 use sqlx::FromRow;
 use sqlx::PgPool;
 
@@ -17,7 +18,7 @@ pub struct UsdtWallet {
     pub network: String,
     pub name: Option<String>,
     pub is_active: bool,
-    pub total_received: Option<f64>,
+    pub total_received: Option<Decimal>,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
